+++ deepxube/base/domain.py
@@ -113,23 +113,25 @@ class Domain(ABC, Generic[S, A, G]):
         """
         pass
 
-    def sample_next_state(self, states: List[S]) -> Tuple[List[S], List[float]]:
+    def sample_next_state(self, states: List[S]) -> Tuple[List[S], List[A], List[float]]:
         """ Get random next state and transition cost given the current state
 
         :param states: List of states
-        :return: Next states, transition costs
+        :return: Next states, actions taken, transition costs
         """
         actions_rand: List[A] = self.sample_state_action(states)
-        return self.next_state(states, actions_rand)
+        states_next, tcs = self.next_state(states, actions_rand)
+        return states_next, actions_rand, tcs
 
-    def random_walk(self, states: List[S], num_steps_l: List[int]) -> Tuple[List[S], List[float]]:
+    def random_walk(self, states: List[S], num_steps_l: List[int]) -> Tuple[List[S], List[List[A]], List[float]]:
         """ Perform a random walk on the given states for the given number of steps
 
         :param states: List of states
         :param num_steps_l: number of steps to take for each state
-        :return: The resulting state and the path cost for each random walk
+        :return: The resulting state, actions taken, and the path cost for each random walk
         """
         states_walk: List[S] = [state for state in states]
+        actions_l: List[List[A]] = [[] for _ in states]
         path_costs: List[float] = [0.0 for _ in states]
 
         num_steps: NDArray[np.int_] = np.array(num_steps_l)
@@ -139,18 +141,19 @@ class Domain(ABC, Generic[S, A, G]):
             idxs: NDArray[np.int_] = np.where(steps_lt)[0]
             states_to_move = [states_walk[idx] for idx in idxs]
 
-            states_moved, tcs = self.sample_next_state(states_to_move)
+            states_moved, actions, tcs = self.sample_next_state(states_to_move)
 
             idx: int
             for move_idx, idx in enumerate(idxs):
                 states_walk[idx] = states_moved[move_idx]
+                actions_l[idx].append(actions[move_idx])
                 path_costs[idx] += tcs[move_idx]
 
             num_steps_curr[idxs] = num_steps_curr[idxs] + 1
 
             steps_lt[idxs] = num_steps_curr[idxs] < num_steps[idxs]
 
-        return states_walk, path_costs
+        return states_walk, actions_l, path_costs
 
     def get_nnet_pars(self) -> List[Tuple[str, str, NNetPar]]:
         return self.nnet_pars
@@ -204,12 +207,11 @@ class ActsRev(Domain[S, A, G], ABC):
 
     """
     @abstractmethod
-    def rev_action(self, states: List[S], actions: List[A]) -> List[A]:
-        """ Get the reverse of the given action
+    def sample_rev_state(self, states: List[S]) -> Tuple[List[S], List[A], List[float]]:
+        """ Get random reverse state, reverse action that returns reverse state to given state and transition cost of that reverse action
 
         :param states: List of states
-        :param actions: List of actions
-        :return: Reverse of given action in the given state
+        :return: Reverse states, reverse actions, reverse transition costs
         """
         pass
 
@@ -280,6 +282,19 @@ class ActsEnumFixed(ActsEnum[S, A, G], ActsFixed[S, A, G]):
 
     def get_num_acts(self) -> int:
         return len(self.get_actions_fixed())
+
+
+# supervised data generation
+class NodesSupervisable(Domain[S, A, G]):
+    @abstractmethod
+    def get_nodes_and_labels(self, steps_gen: List[int]) -> Tuple[List[S], List[G], List[float]]:
+        pass
+
+
+class EdgesSupervisable(Domain[S, A, G]):
+    @abstractmethod
+    def get_edges_and_labels(self, steps_gen: List[int]) -> Tuple[List[S], List[G], List[A], List[float]]:
+        pass
 
 
 # Goal mixins
@@ -366,7 +381,7 @@ class GoalSampGoalStateSamp(GoalStateGoalPairSampleable[S, A, G], GoalSampleable
 
 
 # Problem instance generation mixins
-class StartGoalWalkable(GoalSampleableFromState[S, A, G]):
+class StartGoalWalkable(GoalSampleableFromState[S, A, G], NodesSupervisable[S, A, G], EdgesSupervisable[S, A, G]):
     """ Can sample start states, take actions to obtain another state, and sample a goal from that state"""
     @abstractmethod
     def sample_start_states(self, num_states: int) -> List[S]:
@@ -400,6 +415,35 @@ class StartGoalWalkable(GoalSampleableFromState[S, A, G]):
 
         return states_start, goals
 
+    def get_nodes_and_labels(self, steps_gen: List[int]) -> Tuple[List[S], List[G], List[float]]:
+        states_start: List[S] = self.sample_start_states(len(steps_gen))
+        states_goal, _, path_costs = self.random_walk(states_start, steps_gen)
+        goals: List[G] = self.sample_goal_from_state(states_start, states_goal)
+
+        return states_start, goals, path_costs
+
+    def get_edges_and_labels(self, steps_gen: List[int]) -> Tuple[List[S], List[G], List[A], List[float]]:
+        # start states
+        states_start: List[S] = self.sample_start_states(len(steps_gen))
+
+        # first step
+        states_start_1step, actions_init, tcs_step1 = self.sample_next_state(states_start)
+
+        # account for step_gen == 0
+        for idx in np.where(np.array(steps_gen) == 0)[0]:
+            states_start_1step[idx] = states_start[idx]
+            tcs_step1[idx] = 0.0
+
+        # random walk
+        steps_gen_minus_1: List[int] = np.maximum(np.array(steps_gen) - 1, 0).tolist()
+        states_goal, _, path_costs = self.random_walk(states_start_1step, steps_gen_minus_1)
+        path_costs = (np.array(tcs_step1) + np.array(path_costs)).tolist()
+
+        # sample goal
+        goals: List[G] = self.sample_goal_from_state(states_start, states_goal)
+
+        return states_start, goals, actions_init, path_costs
+
 
 class GoalStartRevWalkable(GoalStateGoalPairSampleable[S, A, G]):
     def sample_problem_instances(self, num_steps_l: List[int], times: Optional[Times] = None) -> Tuple[List[S], List[G]]:
@@ -414,13 +458,13 @@ class GoalStartRevWalkable(GoalStateGoalPairSampleable[S, A, G]):
 
         # random walk to get start states
         start_time = time.time()
-        states_start: List[S] = self.random_walk_rev(states_goal, num_steps_l)
+        states_start: List[S] = self.random_walk_rev_no_path_cost(states_goal, num_steps_l)
         times.record_time("random_walk", time.time() - start_time)
 
         return states_start, goals
 
     @abstractmethod
-    def random_walk_rev(self, states: List[S], num_steps_l: List[int]) -> List[S]:
+    def random_walk_rev_no_path_cost(self, states: List[S], num_steps_l: List[int]) -> List[S]:
         """ Domain need not be reversible as the distance of a path obtained by a number reverse steps can be roughly correlated with the number of steps
 
         :param states: List of states
@@ -430,9 +474,61 @@ class GoalStartRevWalkable(GoalStateGoalPairSampleable[S, A, G]):
         pass
 
 
-class GoalStartRevWalkableActsRev(GoalStartRevWalkable[S, A, G], ActsRev[S, A, G], ABC):
-    def random_walk_rev(self, states: List[S], num_steps_l: List[int]) -> List[S]:
+class GoalStartRevWalkableActsRev(GoalStartRevWalkable[S, A, G], ActsRev[S, A, G], NodesSupervisable[S, A, G], EdgesSupervisable[S, A, G], ABC):
+    def random_walk_rev_no_path_cost(self, states: List[S], num_steps_l: List[int]) -> List[S]:
         return self.random_walk(states, num_steps_l)[0]
+
+    def random_walk_rev(self, states: List[S], num_steps_l: List[int]) -> Tuple[List[S], List[List[A]], List[float]]:
+        states_walk: List[S] = [state for state in states]
+        actions_rev_l: List[List[A]] = [[] for _ in states]
+        path_costs: List[float] = [0.0 for _ in states]
+
+        num_steps: NDArray[np.int_] = np.array(num_steps_l)
+        num_steps_curr: NDArray[np.int_] = np.zeros(len(states), dtype=int)
+        steps_lt: NDArray[np.bool_] = num_steps_curr < num_steps
+        while np.any(steps_lt):
+            idxs: NDArray[np.int_] = np.where(steps_lt)[0]
+            states_to_move = [states_walk[idx] for idx in idxs]
+
+            states_moved, actions_rev, tcs = self.sample_rev_state(states_to_move)
+
+            idx: int
+            for move_idx, idx in enumerate(idxs):
+                states_walk[idx] = states_moved[move_idx]
+                actions_rev_l[idx].append(actions_rev[move_idx])
+                path_costs[idx] += tcs[move_idx]
+
+            num_steps_curr[idxs] = num_steps_curr[idxs] + 1
+
+            steps_lt[idxs] = num_steps_curr[idxs] < num_steps[idxs]
+
+        return states_walk, actions_rev_l, path_costs
+
+    def get_nodes_and_labels(self, steps_gen: List[int]) -> Tuple[List[S], List[G], List[float]]:
+        states_goal, goals = self.sample_goalstate_goal_pairs(len(steps_gen))
+        states_start, _, path_costs = self.random_walk_rev(states_goal, steps_gen)
+
+        return states_start, goals, path_costs
+
+    def get_edges_and_labels(self, steps_gen: List[int]) -> Tuple[List[S], List[G], List[A], List[float]]:
+        # samp_goal_state_goal
+        states_goal, goals = self.sample_goalstate_goal_pairs(len(steps_gen))
+
+        # random walk rev
+        steps_gen_min_1: List[int] = np.maximum(np.array(steps_gen) - 1, 0).tolist()
+        states_start_1step, _, path_costs = self.random_walk_rev(states_goal, steps_gen_min_1)
+
+        # first step
+        states_start, actions_init, tcs_step1 = self.sample_rev_state(states_start_1step)
+
+        # account for step_gen == 0
+        for idx in np.where(np.array(steps_gen) == 0)[0]:
+            states_start[idx] = states_start_1step[idx]
+            tcs_step1[idx] = 0.0
+
+        path_costs = (np.array(tcs_step1) + np.array(path_costs)).tolist()
+
+        return states_start, goals, actions_init, path_costs
 
 
 # numpy convenience mixins
@@ -444,8 +540,9 @@ class NextStateNP(Domain[S, A, G]):
 
         return states_next, tcs
 
-    def random_walk(self, states: List[S], num_steps_l: List[int]) -> Tuple[List[S], List[float]]:
-        states_np = self._states_to_np(states)
+    def random_walk(self, states: List[S], num_steps_l: List[int]) -> Tuple[List[S], List[List[A]], List[float]]:
+        states_np: List[NDArray] = self._states_to_np(states)
+        actions_l: List[List[A]] = [[] for _ in states]
         path_costs: List[float] = [0.0 for _ in states]
 
         num_steps: NDArray[np.int_] = np.array(num_steps_l)
@@ -460,15 +557,17 @@ class NextStateNP(Domain[S, A, G]):
 
             for l_idx in range(len(states_np)):
                 states_np[l_idx][idxs] = states_moved[l_idx]
+
             idx: int
             for act_idx, idx in enumerate(idxs):
+                actions_l[idx].append(actions_rand[act_idx])
                 path_costs[idx] += tcs[act_idx]
 
             num_steps_curr[idxs] = num_steps_curr[idxs] + 1
 
             steps_lt[idxs] = num_steps_curr[idxs] < num_steps[idxs]
 
-        return self._np_to_states(states_np), path_costs
+        return self._np_to_states(states_np), actions_l, path_costs
 
     @abstractmethod
     def _states_to_np(self, states: List[S]) -> List[NDArray]: